made the service class and the controller class

+++ src/main/java/com/sum/url_shortener/entity/DBRecord.java
@@ -10,7 +10,6 @@ import jakarta.persistence.Table;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
 
 @Entity
 @Table(name = "urls",
@@ -26,7 +25,7 @@ public class DBRecord {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Column(name = "long_url")
     private String longUrl;
